+++ Tests/DatasourceTestCase.swift
@@ -378,5 +378,22 @@ class DatasourceTestCase: XCTestCase {
         let _ = controller.manager.tableDelegate?.tableView(controller.tableView, viewForFooterInSection: 0)
         waitForExpectations(timeout: 0.1)
     }
+    
+    func testNoCellMappingsTriggerAnomaly() {
+        let exp = expectation(description: "No cell mappings found for model")
+        let anomaly = DTTableViewManagerAnomaly.noCellMappingFound(modelDescription: "3")
+        controller.manager.anomalyHandler.anomalyAction = exp.expect(anomaly: anomaly)
+        controller.manager.memoryStorage.addItem("3")
+        #if os(tvOS)
+        let _ = controller.manager.tableDataSource?.tableView(controller.tableView, cellForRowAt: indexPath(0, 0))
+        #endif
+        if #available(iOS 11, *) {}
+        else {
+            let _ = controller.manager.tableDataSource?.tableView(controller.tableView, cellForRowAt: indexPath(0, 0))
+        }
+        waitForExpectations(timeout: 0.1)
+        
+        XCTAssertEqual(anomaly.debugDescription, "❗️[DTTableViewManager] UITableView requested a cell for model, but view model mapping for it was not found, model description: 3")
+    }
 #endif
 }